cleanup: move Placholder into placeholder package
