+++ src/main/java/net/kyori/adventure/text/minimessage/parser/TokenParser.java
@@ -32,8 +32,6 @@ import java.util.function.BiPredicate;
 import java.util.function.Function;
 import java.util.function.IntPredicate;
 import java.util.function.Predicate;
-import net.kyori.adventure.text.minimessage.Placeholder;
-import net.kyori.adventure.text.minimessage.Placeholder.StringPlaceholder;
 import net.kyori.adventure.text.minimessage.Tokens;
 import net.kyori.adventure.text.minimessage.parser.match.MatchedTokenConsumer;
 import net.kyori.adventure.text.minimessage.parser.match.StringResolvingMatchedTokenConsumer;
@@ -44,6 +42,8 @@ import net.kyori.adventure.text.minimessage.parser.node.RootNode;
 import net.kyori.adventure.text.minimessage.parser.node.TagNode;
 import net.kyori.adventure.text.minimessage.parser.node.TagPart;
 import net.kyori.adventure.text.minimessage.parser.node.TextNode;
+import net.kyori.adventure.text.minimessage.placeholder.Placeholder;
+import net.kyori.adventure.text.minimessage.placeholder.Placeholder.StringPlaceholder;
 import net.kyori.adventure.text.minimessage.placeholder.PlaceholderResolver;
 import net.kyori.adventure.text.minimessage.transformation.Inserting;
 import net.kyori.adventure.text.minimessage.transformation.Transformation;
