+++ src/main/java/net/kyori/adventure/text/minimessage/placeholder/DynamicPlaceholderResolver.java
@@ -25,7 +25,6 @@ package net.kyori.adventure.text.minimessage.placeholder;
 
 import java.util.function.Function;
 import net.kyori.adventure.text.ComponentLike;
-import net.kyori.adventure.text.minimessage.Placeholder;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
