+++ src/main/java/net/kyori/adventure/text/minimessage/placeholder/EmptyPlaceholderResolver.java
@@ -23,7 +23,6 @@
  */
 package net.kyori.adventure.text.minimessage.placeholder;
 
-import net.kyori.adventure.text.minimessage.Placeholder;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
