+++ src/main/java/net/kyori/adventure/text/minimessage/placeholder/FilteringPlaceholderResolver.java
@@ -24,7 +24,6 @@
 package net.kyori.adventure.text.minimessage.placeholder;
 
 import java.util.function.Predicate;
-import net.kyori.adventure.text.minimessage.Placeholder;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
