+++ src/main/java/net/kyori/adventure/text/minimessage/placeholder/PlaceholderResolver.java
@@ -32,7 +32,6 @@ import java.util.Objects;
 import java.util.function.Function;
 import java.util.function.Predicate;
 import net.kyori.adventure.text.ComponentLike;
-import net.kyori.adventure.text.minimessage.Placeholder;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
