+++ src/main/java/net/kyori/adventure/text/minimessage/transformation/TransformationRegistryImpl.java
@@ -27,10 +27,10 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import net.kyori.adventure.text.minimessage.Context;
-import net.kyori.adventure.text.minimessage.Placeholder;
-import net.kyori.adventure.text.minimessage.Placeholder.ComponentPlaceholder;
 import net.kyori.adventure.text.minimessage.parser.ParsingException;
 import net.kyori.adventure.text.minimessage.parser.node.TagPart;
+import net.kyori.adventure.text.minimessage.placeholder.Placeholder;
+import net.kyori.adventure.text.minimessage.placeholder.Placeholder.ComponentPlaceholder;
 import net.kyori.adventure.text.minimessage.placeholder.PlaceholderResolver;
 import net.kyori.adventure.text.minimessage.transformation.inbuild.PlaceholderTransformation;
 import org.jetbrains.annotations.NotNull;
