+++ src/main/java/net/kyori/adventure/text/minimessage/transformation/inbuild/PlaceholderTransformation.java
@@ -26,7 +26,7 @@ package net.kyori.adventure.text.minimessage.transformation.inbuild;
 import java.util.Objects;
 import java.util.stream.Stream;
 import net.kyori.adventure.text.Component;
-import net.kyori.adventure.text.minimessage.Placeholder;
+import net.kyori.adventure.text.minimessage.placeholder.Placeholder;
 import net.kyori.adventure.text.minimessage.transformation.Inserting;
 import net.kyori.adventure.text.minimessage.transformation.Transformation;
 import net.kyori.adventure.text.minimessage.transformation.TransformationFactory;
